#1 refactor into use container hosting
