--- templates/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-{% with messages = get_flashed_messages() %}
-  {% if messages %}
-    <ul class=flashes>
-    {% for message in messages %}
-      <li>{{ message }}</li>
-    {% endfor %}
-    </ul>
-  {% endif %}
-
-{% endwith %}
-
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>File Upload Form</title>
-  </head>
-  <body>
-    <h1>File Upload Form</h1>
-    <form action="{{ url_for('upload') }}" method="POST" enctype="multipart/form-data">
-      <input type="file" name="file">
-      <button type="submit">Upload</button>
-    </form>
-  </body>
-</html>
-
